MealPrep360-ImageService/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Image Processing
Pillow==10.2.0
imageio==2.33.1
opencv-python-headless==4.9.0.80

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6

# Logging
loguru==0.7.2

# Redis
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

